fix(dispatch): Simulation results not saved for some simulations and overall status doesn't reflect such errors (#2428)
The  simulation is now marked as failed if processing/uploading the results fails.
All results and logs are now updated independently. The issue with  results not being uploaded if logs failed was caused by the fail fast behavior of promise.All(). Using the new es2020 promise.allSetteled() fixes this problem, but requires updating the base tsconfig.

Fixes #2416