fix(simulators-api): Allow for date based versions
The api using a version-comparision library to determine what the latest version of each tool is.
The library depends on versions being in semantic format. However, there are some tools that have date based versions such as "2020-02-15". This change replaces the "-" charecters with "." during the comparision to enable correct sorting.

fix #2681