refactor: consolidate backend apis
The ontology, dispatch, and platform apis are consolidated into one main backend api for
biosimulations

BREAKING CHANGE: The ontology, dispatch, and platform apis are consolidated into one main backend
api for biosimulations. There is a seperate api for biosimulators. The combine-service also provides a rest api that is mostly intended for internal use.

closes #2724