revert: revert change to build front end apps
The --prod coniguration also includes build optimizations that are not included i the dev
configuration. This means we need to either replace the files manually, or set the correct build
target in the app config, which is more involved.We also want to use stage to keep this file from
any changes that can be made to dev locally.