fix(dispatch): fix displaying of data in cases where simulation fails and metadata is not present
if the run failed and did not contain any metadata, then the view service returned undefined. This
was indistingushable from cases where the metadata had not yet loaded, causing the view to display a
spinner. Instead, it now returns null, which lets the view know that the data is loaded, it is just
not present,and display the appropriate error on the metadata tab

fix #3705