perf(api): extract the combine archive directly on s3 via streams
uses streams to directly extract the combine archive onto s3 as soon as the archive is uploaded.
This removes the "extract" step from the slurm workflow, and is significantly faster. This also
opens up the possibility of the combine-api responding to an event to access the files that it needs
rather than waiting for the post processing to send it the files from the simulation run.

closes #3094