refactor(api): simplify management of files
No metadata about files is managed directly. Instead a simple url is added to the datamodel to keep
track of the location of the simulation run input archive. This url will always point to our s3 bucket since files posted directly to the API are uploaded to s3, and a copy of any archives provided via url is saved directly to s3.
The previous handling of the metadta of the uploaded file had not been needed since switching to
exclusively using s3, but the database structure still included metadata about files.

BREAKING CHANGE: the download project endpoint will fail for all previously submitted simulation
runs. Runs submitted prior to this change will not be retrievable by the api or applications