fix(dispatch-service): fix ssh connection for stale connections
when the dispatch service did not use the ssh connection for long, it would error the next time it
was used. this was adressed partially by improving the logic for reconnecting on connection loss.
However, the command will still fail. This is adressed in commmits that add a queue with retying for
most operations that need ssh connections

closes #4099