perf(dispatch-service): increase concurrency of proccessing steps
set the concurrency of all the processors to 10. it seems that most of the processing steps are
network bound, with the possible exception of the extraction and thuimbnail processing. It may be
neccessaty to tweak the concurrency for those steps if they are using a lot of cpu. There may also
be more room to increase the concurrency for some of the steps

closes #4047