feat(dispatch): apply new style to the create project form
Apply a multi-step flow to the create project form. Logic for translating objects from the format
they're returned from combine, to the format used within dispatch and the forms, and then back is
factored out into an ArchiveCreationUtility that has static helper functions to perform this work.
The UI of the create project form is broken into several form-step components that are displayed one
at a time as the user proceeds through the form.