refactor(ui): begin to remove table component dependency from controls
the controls of the table depend on recieving the table as input. the logic for the actual
datamanipulation is within the table component. The controls component calls these methods based on
UI changes. However, the controls unit is more general that the table, and can create events
instead, so that any component can import and use the controller for its own internal logic.