feat(dispatch): do not display the archive selection or capabilities form steps when re-running
When re-running a simulation run via the runs/new endpoint, if a valid project url is provided and
the archive can be fetched then do not display the form step for providing an archive url or file.
Likewise, if a valid simulator is set in the params, don't display the capabilities form step which
is used to help select a simulator.