refactor: update styling to latest
angular material has been updated with many changes to the sass functions and mixins. Our theme system has drifted overtime and become rather non standard, so these updates were not caught by the auto migrations

Manually refactored and updated to the latest versions. This is a good target for refactoring and simplifying.